backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pgvector==0.3.6

# Pydantic
pydantic==2.10.3
pydantic-settings==2.6.1

# Auth (JWT verification)
python-jose[cryptography]==3.3.0
httpx==0.27.2

# File Parsing
pdfminer.six==20231228
python-docx==1.1.2

# AI
google-generativeai==0.8.3
groq>=1.0.0

# Payments
stripe==11.2.0

# NLP / Scoring
scikit-learn==1.6.0
numpy==2.1.3

# Security
slowapi==0.1.9
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
